requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
fastapi==0.85.1
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.0.0
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
Unidecode==1.3.6
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.3
